monthly-analysis-function/requirements.txt
azure-functions>=1.18.0,<2.0.0
azure-keyvault-secrets>=4.8.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
openai>=1.35.0
requests>=2.31.0